+++ circuit_maintenance_parser/providers.py
@@ -1,5 +1,6 @@
 """Definition of Provider class as the entry point to the library."""
 import logging
+import traceback
 from typing import Iterable, Type
 
 from pydantic import BaseModel, Extra
@@ -64,7 +65,11 @@ class GenericProvider(BaseModel, extra=Extra.forbid):
             data_type: Hint to limit the parsing to specific data types. (default "")
 
         """
+        error_message = ""
+        provider_name = self.__class__.__name__
+        related_exceptions = []
         for parser_class in self._parser_classes:
+            parser_name = parser_class.__name__
             try:
                 parser = parser_class(
                     raw=self.raw,
@@ -73,12 +78,20 @@ class GenericProvider(BaseModel, extra=Extra.forbid):
                 )
                 if data_type and data_type == parser.get_data_type() or not data_type:
                     return parser.process()
-            except (ParsingError, MissingMandatoryFields):
+            except (ParsingError, MissingMandatoryFields) as exc:
                 logger.debug(
-                    "Parser %s for provider %s was not successful", parser_class.__name__, self.__class__.__name__
+                    "Parser %s for provider %s was not successful:\n%s",
+                    parser_name,
+                    provider_name,
+                    traceback.format_exc(),
                 )
+                error_message += f"- Parser class {parser_name} from {provider_name} failed due to: {exc.__cause__}\n"
+                related_exceptions.append(exc)
                 continue
-        raise ParsingError("None of the parsers was able to parse the notification")
+        raise ParsingError(
+            f"None of the {provider_name} parsers was able to parse the notification.\nDetails:\n{error_message}",
+            related_exceptions=related_exceptions,
+        )
 
 
 ####################
@@ -123,6 +136,13 @@ class PacketFabric(GenericProvider):
     """PacketFabric provider custom class."""
 
     _default_organizer = "support@packetfabric.com"
+
+
+class Telia(GenericProvider):
+    """Telia provider custom class."""
+
+    _parser_classes: Iterable[Type[Parser]] = [ICal]
+    _default_organizer = "carrier-csc@teliacompany.com"
 
 
 class Telstra(GenericProvider):