+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.2.2 - 2021-08-12
+
+### Added
+
+- #41 - Added new provider `Telia` using `Ical`
+- #42 - Improve error and debug messages to ease troubleshooting of parsing issues
+
 ## v1.2.1 - 2021-06-22
 
 ### Fix